feat: 添加axios依赖并优化实体类字段类型
refactor(backend): 将实体类中的Integer/Boolean字段改为Byte/TINYINT类型
build(frontend): 添加axios依赖用于前端HTTP请求

+++ backend/src/main/java/com/yangshengzhou/backend/service/AuthService.java
@@ -87,7 +87,7 @@ public class AuthService {
         user.setUsername(username);
         user.setEmail(email);
         user.setPassword(passwordEncoder.encode(password));
-        user.setRole(0); // 使用整数类型
+        user.setRole((byte) 0); // 使用byte类型
         user.setCreatedTime(LocalDateTime.now());
         user.setUpdatedTime(LocalDateTime.now());
